+++ src/index.mjs
@@ -1,5 +1,5 @@
 export { data, isData } from './data.mjs';
-export { trait, isTrait, _ } from './trait.mjs';
+export { apply, trait, isTrait, _ } from './trait.mjs';
 export { complect } from './complect.mjs';
 //export { memoFix } from './memoFix.mjs';